test: Fix oo.EventEmitter context test failure on nodejs
Contrary to in browsers, the global object in node is not
shared between scopes. Each "require" gets its own global
object so we can't compare them by reference. Fallback to
asserting that both 'this' and 'global' are node-globals by
checking non-emptyness of the '$ref' property. It doesn't
really matter what the global looks like, it just needs to
be the node global (as opposed to null, undefined, the
EventeEitter instance, ..)