test: Fix false positive in nodejs EventEmitter test
Node has no property '$ref' on the global object, this is
just how some debuggers indicate a circular reference.

The premise was that the expected value would be 'true' (I
should've harcoded that) as it would always be on the global
one and we're asserting whether the bound context has it as
well.

I thought the global objects are unique per module in node
because this !== global, however I was wrong about that too.
Node's global object are indeed truly global accross modules
(though usage of this feature is obviously discouraged).

The reason `this !== global` is because (unlike in browsers)
in node the global object is not bound to `this` in the global
scope. Instead `this` in the global scope is bound to
`module.exports`. To get node's global object you need to be
inside a closure with no overriden call context and there
`this` will be bound to the global object.

It being hard to access signifies that it shouldn't be used,
we're merely referencing it here in the tests to assert
behaviour, we aren't (and shouldn't) be using in the actual
library.