inheritClass: Use Class.super instead of this.constructor.super
Using `this.constructor.super` is problematic when dealing with
any class that has a subclass as it will yield the parent of the
child (e.g. the current class again) because 'this' will be an
instance of the child class, not the class it is defined in.

The documented example actually ran into an infinite recursion
with "Uncaught RangeError: Maximum call stack size exceeded".

Also changed oo -> OO as this example has no local 'oo' alias.

Bug: 62405
Change-Id: Ia17e64482509a68e13a1ae4639c1eb9d090073e5