core: Use bracket notation for 'super' for ES3 compatibility
The 'super' will remain for compatibility and for code targetting
only newer browsers, but inside oojs we assign it using bracket
notation so that it can be parsed by ES3 engines.

The 'parent' alias is for code needing compatibuility with older
browsers (instead of requiring them to use ['super'] which is
undesirable by coding style).

Bug: 63303
Change-Id: I2e84dc592b3d6adb5d17989380523e007d2ada05