test: Fix broken getHash test for iframe Object
This test was guarded by a condition so that it doesn't run in
nodejs. However this condition was false even in browsers:

> doc = this && 'window' in this && this.window.document;

The 'this' here is the QUnit test environment object.

Convert this and the iframe code to match the pattern we use
elsewhere (e.g. util.test.js) using global.document.

Follows-up 39d751d.

Change-Id: I4920c5a72e7ad53cb27d14ee6a516fa47ac577d0