EventEmitter: Remove redundant 'if' statement in off()
This was added when we fixed a bug in disconnect() where it
threw due to accessing an undefined property if there were no
events bound yet.

However, off(), unlike disconnect(), never had this bug. The 'if'
statement is redundant as the 'delete' operator checks this
already. We don't use its return value either (which is true/false
for whether it deleted something).

Increases code coverage by removing implicit 'else' branches that
were never ran.

Change-Id: I40cba9292ad851757fc2a605f0ddc66df0ea173f