build: Update to karma-qunit v0.1.4 and qunitjs v1.15
A bug in qunitjs v1.14 caused karma-qunit to report "Executed 31 of 0"
tests. karma-qunit v0.1.3, however, hardcoded a dependency on qunitjs
v1.14, making it impossible to upgrade to qunitjs v1.15 (which fixed
the bug).

karma-qunit v0.1.4 laxes its peer dependency on qunitjs to "^1.14.0"
allowing its users to use a newer version of QUnit.

Changes
* https://github.com/karma-runner/karma-qunit/commits/v0.1.4
* https://github.com/jquery/qunit/blob/1.15.0/History.md

Change-Id: Ib4f4be5679191bfba57d6dfa2aa8446525c79088