core: Remove dependency on Object.create
We claim not to use any unpolyfillable features of ES5, so let's not
depend on this method. We only depend on a subset of its features that can
be mostly polyfilled, so we were technically not lying, but it's just so
much simpler this way.

We now include a fallback implementing the minimal set of required features.
It's just a few lines, a small price to pay for the convenience.

It also has the benefit of not requiring users to assign
Object.create with a partial implementation, which can confuse
other libraries.

Bug: T109985
Change-Id: I63e23c7b56cc8003293345fb08ae02d7a9c4de54