EmitterList: Change insertItem from @private to @protected
This method was always supposed to be an extendable method, especially
when taking into account dealing with widgets (such as GroupWidget)
which uses the EmitterList functionality but extends it to interact
with the DOM. The tag @private is therefore wrong and should be
replaced with @protected.

Further explanation (and a warning) was added to the documentation
to clarify the use and extension of the method.

For an example of usage, see I7b9d959eff563

Bug: T114707
Change-Id: I2e289bdc59246feea16e1a9b224014959ebc9eec