core: Improve error message when inheritClass/mixinClass called with 'undefined'
Currently, calling `OO.inheritClass( MyClass, OO.Foo )` (where OO.Foo
is undefined because you forgot to specify a dependency) results in a
fairly confusing exception like "Right-hand side of 'instanceof' is
not an object". I've debugged issues like this at least two times in
the past so I think the extra check is worth it.

Bug: T145315
Change-Id: I317ddaa50245aeed4a8eaf299dd38808fdd357e0