tests: Fix broken setup/teardown in tests/index.html
In QUnit 1.16, module setup/teardown was renamed to beforeEach
and afterEach (old names kept for back-compat).

The back-compat aliases were removed in QUnit 2.0.

This didn't break the karma run because we only used setup/teardown
to create an element qunit-fixture for each test, which qunit-karma
does internally for us as well (because that too is a deprecated
feature that was removed long ago).

As such, these setup/teardown handles have been unused for a while
but didn't cause problems unless running tests manually.

Change-Id: I6135706db9c79a2f49279d03fe772bf0a404ff98