build: Disable eslint sourceType=module option
Follows-up 93dccfe, which accidentally enabled this in order to
silence an (otherwise) innocent lint warning.

The "module" source type is reserved for when writing code in
ES6 or newer in the "ES Module" manner, which is an new way of
parsing and executing JavaScript that was inspired by run-time
conventions of closure, require(), and module.exports, and
standardised a syntax that would make that work natively in both
Node and in browsers. However no browser has implemented that yet
and it's still behind an experimental flag in Node.js as well.
It also doesn't make sense in combination with ecmaVersion=5.

The warning introduced by ESLint 4.0.0, that was silenced by this
unrelated option:

> [core.js:9:2]  error   no-implicit-globals:
> Implicit global variable, assign as global property instead.

The documentation mentions that this is for detecting an issue
that doesn't affect CommonJS (e.g. Node) and ESModules, given
those run-times use local scope by default.

In our case, we concatenate intro.txt and outro.txt during the
build, which actually result in this file being in a closure
and thus not create global variables.

There isn't really a way to tell ESLint about our implied local
scope, so I'll just disable this rule for now. This is essentially
the same as it was with sourceType=module, given the rule
effectively already disables itself in that mode. But the
difference is that disabling 'no-implicit-globals' only doesn't
have other side effects like tolerating all other ESModule-mode
syntax and parser features, which we definitely don't want to
allow in this file.

Change-Id: Ib65f576537390b4905b921de27bfbb818ca3f56c