build: Add AUTHORS.txt back to package root, remove dupes from in dist/
The previous commit was relying on the automatic whitelisting of text files,
which covers README (through npm's default mandatory inclusion of of *.md,
LICENSE* and package.json), but not AUTHORS.txt.

This commit fixes that by making inclusion of README.md explicit, and
also adding AUTHORS.txt back.

The v2.2.1 release still contained a copy of AUTHORS.txt in dist/ however.
This commit also removes the duplication of these text files in dist/, given
that otherwise there would be two copies of "LICENSE*" and "*.md" in the
published package.

See also I4bd91811 (mediawiki/core) which updates update-oojs.sh to
copy the text files from the root of the package instead of dist/.

Change-Id: Ibcbb73961a2692141409f1636973e248d30e2087