Use "OO" internally too instead of "oo"
Using "oo" internally was causing issues when generating
documentation using JSDoc3. Each class, method, property,
and event had to be specifically named with OO.* to work
properly. Easier to just use "OO" internally too.

This doesn't make any changes to the tests which passes
"OO" as an argument and uses it as "oo".

Change-Id: Ib32d850dedf108b7392f5ab65c50bb23341697a6