build: Honour ENV CHROMIUM_FLAGS for Wikimedia CI
Travis CI has switched away from Docker and now runs Headless Chrome
out of the box in its VMs without needing --no-sandbox.

Meanwhile, Wikimedia CI has switched *to* Docker and thus requires
--no-sandbox (passed by CI's env as CHROMIUM_FLAGS to avoid having
to hardcode it specifically, in case we find a way to fix it later).

Update Travis CI config to test on Node 10, remove the deprecated
no-op flag "sudo: false", and remove the redundant now-default of
"dist: trusty" (if and when the default changes, we want to change
along, as this doesn't actually do anything Ubuntu specific).

Also:
* Firefox was still non-headless, which I've switched to headless so
  that we can make use of the lighter Jenkins job without Xvfb.

* Bump SauceLabs testing of Safari old from 8 to 9, as they've
  stopped providing testing on Safari 8.

* Lower SauceLabs testing of Sasfari latest from 12 to 11,
  as their Safari 12 server are currently having a broken version
  of SauceLabs Connect Tunnel, where every request leads to HTTP
  500.

Change-Id: I527a5ad5927e630ccaba412a0fa0cbe9d5d21c93