build Revert package.json Git repo url
This reverts commit 2161b27be00c3296114fd65152bda228f24e23fd.

This URL is required to be a valid Git remote, not a web page.
See also <https://docs.npmjs.com/files/package.json#repository>

For GitHub, it supports HTTPS and Git+HTTPS over the same path, but
for Gerrit and most other code hosts or other version control systems
this is not the case.

Change-Id: Id8a8cbb55d9a5919800b331ff192bcce49027077