[BREAKING CHANGE] build: Remove special "jquery" build
I added this originally in 2013 to ensure OOjs remains standalone, and
thus decoupled from jQuery. This was a good call, I think. The idea was
that if the project grew, it would encourage the library to remain
standalone and introduce needless dependencies on complex jQuery methods.

While the project did grow, it was in a separate project (OOUI) and
there I did not remember to do the same, and so we never did.

For OOjs, though, I think we can start to take the next step and just
offer the standalone distribution only, it does not seem worth while the
handful of bytes this saves, given the the build, test, and release
complexity it involves. And, by doing this, we can remove the "jquery"
dependency downstream in MediaWiki and thus push that a little further
down the chain. (Even though it was never a hard dependency, as
demonstrated by this build.)

Change-Id: I340867f38a0dd2103eb43ee8e633bd11ee5da4ea