Factory: Add support for registering by `Class.key`
== Warning

We will be discussing a static property named "static",
and a name named "name".

== Why

Using `Class.static.name` is impractical in combination with ES6
classes for two reasons.

1. Inheritence would work incorrectly.

The `Class.static` subobject would, in an ES6 class, refer directly and
in its entirety to the parent object, instead of inheriting individual
properties.

In OOjs, we setup inheritance from `Foo.prototype` and `Foo.static`,
to `SubFoo.prototype` and `SubFoo.static`.

In ES6, when using the `extends` keywords only (and not calling
`OO.initClass`), there is inheritence from `Foo.prototype` and `Foo`
itself, to `SubFoo.prototype` and `SubFoo`. This means `Foo.static`
would be itself just another static property that happens to be named
"static", `Foo.static === SubFoo.static`, and thus SubFoo.static.x=1
would affect Foo as well, which is wrong. This is effectively the same
as what we can do today if you had something like
`Foo.static.list = [x]` and then do something like
`SubFoo.static.list.push(y)`. This would also modify the original array
but is less surprising, as one is meant to set static members and
not modify by-reference, and is as such avoided when setting
`SubFoo.static.list = [x, y]`.

2. Conflict with reserved property called "name"

In ES6 classes, static members are own properties on the class
constructor, which is technically a function. Functions already have
a reserved `name` property. Hence, we can't simply migrate from
`Class.static.name` to `Class.name`, but we have to actually choose
a different name for the "name" concept of OOjs Factory.

On the task, there is consensus for calling this a "key".

This patch updates documentation to refer to this as "key", and add
support for registring by `Class.key`, and keeping back-compat for
`Class.static.name`.

Bug: T96640
Change-Id: If29228e7eff8842838af6bbd63fa10574423c30e