Factory: Support registering and creating objects from ES6 classes
ES6 classes are generally understood as syntax sugar, whereby
"MyClass" becomes the constructor function, and the other fields
and methods are assigned to this function object as own properties or
as own properties to "MyClass.prototype", and thus are generally
transparently compatible with plain JS logic such as `Object.create()`,
including the ability to create an ES6 class that extends a plain JS
class, and vice versa to create a plain JS function constructor that
extends a native ES6 class.

But... while this is all true, the spec has a number of additional
nominal statements, one of which is that the function object of a
class constructor has an implicit hidden check when called that throws
an error if it is called without being in the `new` operator context,
which made it so that if you passed an ES6 class to an OO.Factory,
and then tried to `create()` it, it threw:

```
class Foo{}

Foo.call(Object.create(Foo.prototype)) // like `new Foo()`
Foo.apply(Object.create(Foo.prototype)) // like `new Foo()`

//> Uncaught TypeError: class constructors must be invoked with 'new'
```

The reason we don't use `new` is that `factory.create()` forwards
variadic arguments. Switch to using Reflect.construct() instead of
Object.create() which satisfies the restriction, as it is effectively
also what the the ES6 Proxy API uses for proxying the `new` operator.

Bug: T96640
Bug: T284935
Change-Id: I8da0aaf28cd20746d1e1c82ab6afb98714a7dfc4