tests: Update SauceLabs config to make this pass
Some time in the past year or so these have gone case-sensitive and
TitleCase for Firefox and Edge (but not Chrome, Safari, or MSIE).

Whilst we're at it, specify exactly the implicit 'latest' tag for
Chrome, Firefox, and Edge, the '11' version for MSIE, and update
the macOS version to latest to get a more current version of
Safari.

Change-Id: I8fbf2ebe2f6e7959d215bcdc2a87dfaeb9d6e281